feat: update modal implementation
