--- src/ui/core/modal/index.tsx
@@ -1,3 +1,0 @@
-export * from './dynamic-modal';
-export * from './modal';
-export * from './types';
