--- src/ui/core/modal/modal-backdrop.tsx
@@ -1,7 +1,0 @@
-import type { BottomSheetBackdropProps } from '@gorhom/bottom-sheet';
-import { BottomSheetBackdrop } from '@gorhom/bottom-sheet';
-import React from 'react';
-
-export const renderBackdrop = (props: BottomSheetBackdropProps) => (
-  <BottomSheetBackdrop appearsOnIndex={0} disappearsOnIndex={-1} {...props} />
-);
