--- src/ui/core/modal/x-close.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import type { SvgProps } from 'react-native-svg';
-import Svg, { Path } from 'react-native-svg';
-
-const SvgComponent = ({ fill = '#001136', ...props }: SvgProps) => (
-  <Svg width={24} height={24} fill="none" viewBox="0 0 24 24" {...props}>
-    <Path
-      d="M18.707 6.707a1 1 0 0 0-1.414-1.414L12 10.586 6.707 5.293a1 1 0 0 0-1.414 1.414L10.586 12l-5.293 5.293a1 1 0 1 0 1.414 1.414L12 13.414l5.293 5.293a1 1 0 0 0 1.414-1.414L13.414 12l5.293-5.293Z"
-      fill={fill}
-    />
-  </Svg>
-);
-
-export const XClose = SvgComponent;
